==> scripts/old_version/fromgapp/01a_OBIS-species-list.Rmd <==
---
title: "01a_OBIS-species-list"
author: "Kate Sheridan"
date: "3/29/2023"
output: html_document
---

This script pulls species data from OBIS to generate a species list for a defined region. By ecoregion section uses shapefiles from the MEOW nearshore ecoregions and PPOW pelagic provinces. 

Currently it pulls and cleans a checklist() but can be updated later for occurrence()

OBIS already uses the WoRMS taxonomic backbone so no adjustment to taxonomy was necessary.
Note that currently Species and Subspecies is preserved; in some cases the species column for a subspecies is unique. 
So when using these lists for analysis: `select(accepted_name) %>% distinct()` instead of `filter(taxon_rank == 'species')`

Output includes:
`accepted_name` -- The most up to date name from WoRMS
`verbatim_name` -- column coalesced from subspecies and species, includes unaccepted names.
`taxonomic_status` -- status of `verbatim_species`
`taxon_rank` -- level of identification
taxonomy: species, subspecies, genus, family, order, class, phylum (invertebrates only)
`is_marine`, `is_brackish`, `is_freshwater`: true/false
`taxon_id` -- should be WoRMS AphiaID for lowest found taxon
`genus_id` -- (invertebrates only) to help look for NCBI IDs
`ncbi_id` -- ID for taxon in NCBI
`region` -- region found in, from MEOW or PPOW


```{r setup, include=FALSE}
library(rfishbase)
library(robis)
#library(obistools) # for cleaning occurrence data
library(taxize)
library(tidyverse)
library(here)
# spatial libraries
library(rnaturalearth)
library(sf)

# here filepaths
splists <- here('processeddata', 'species_lists', 'region')
spatial <- here("rawdata", "spatial_data", "MEOW")
```



# By ecoregion

```{r set up ecoregions}
# World map
# basemap
world_map <- ne_countries(scale = 'small', returnclass = c("sf"))

meow <- read_sf(here(spatial, "meow_ecos_simple.shp")) %>% #simplified in QGIS using Simplify function: method = Distance (Douglas-Peuker), Tolerance = 1
  dplyr::select(c('ECOREGION', 'geometry')) %>%
  rename(region = ECOREGION) %>%
  mutate(type = 'meow')
ppow <- read_sf(here(spatial, "ppow_simple_NEP.shp")) %>% #simplified in QGIS using Simplify function: method = Distance (Douglas-Peuker), Tolerance = 1
  dplyr::select(c('PROVINC', 'geometry')) %>%
  rename(region = PROVINC) %>%
  mutate(type = 'ppow')

regions <- rbind(meow,ppow)

# Base map
## what does kk mean?
kk <- ggplot() +
  # fill = col to remove country borders
  geom_sf(data = world_map, size = .2, fill = "gray80", col = "gray80") + 
  theme(panel.grid.major = element_line(color = gray(0.9), 
                                        linetype = "dashed"))

```

```{r checklist function}
# requires robis to be loaded

# takes a dataframe that has our shapefiles loaded in with two main columns:
# wkt from st_as_txt() and region
obis_checklist <- function(region_df) {
  # make blank object
  search <- tibble()
  region_geo <- region_df$wkt
  region_name <- region_df$region
  # loop by region
  for (i in 1:length(region_geo)) {
    #region_geo <- region_df$wkt
    print(paste0('searching for ', region_name[i]))
    found_sp <- checklist(geometry = region_geo[i],
                          startdate = '1945-01-01') %>%
      # only keep animals
      filter(kingdom == 'Animalia') %>%
      # only keep things at the levels we want
      filter(taxonRank %in% c('Species', 'Subspecies',
                              'Tribe', 'Variety', 'Forma',
                              'Genus', 'Subgenus', NA))
      found_sp$region = paste(region_name[i])
    message('done')
    
    search <- bind_rows(search, found_sp)
  }
  return(search)
}
```


```{r meow checklist}
#Ecoregions
# check on map
kk +  
  geom_sf(data = regions, 
          aes(fill = region), 
          linewidth = .2, 
          col = 0, 
          alpha=.3) +
  ggtitle("Ecoregions") +
  theme(legend.position = "none") +
  coord_sf(expand = FALSE)


#meow
regions$wkt <- st_as_text(regions$geometry) #create well known text of ecoregions
# make sure its all valid?
st_is_valid(regions$geometry, reason = TRUE)


regions_nep <- regions %>% #filter for regions of interest
  filter(region %in% c("Aleutian Islands", "Gulf of Alaska", "North American Pacific Fijordland", 
                        "Puget Trough/Georgia Basin", "Oregon, Washington, Vancouver Coast and Shelf",
                        "Northern California", "Southern California Bight", "California Current",
                        "Eastern Tropical Pacific", "North Central Pacific Gyre", "North Pacific Transitional",
                        "Subarctic Pacific"))

# use function to extract species list by region
regions_nep_checklist <- obis_checklist(regions_nep)


```

### Fish

```{r meow-fish}
nep_fish <- regions_nep_checklist %>%
  # only species level records
  filter(taxonRank %in% c('Species', 'Subspecies')) %>%
  # only fish; polyphyletic so can be complicated
  # class covers most of it, some sculpins lack class though
  filter(class %in% c('Actinopteri', 'Elasmobranchii', 'Holocephali',
                      'Petromyzonti', 'Myxini') |
           order %in% c('Scorpaeniformes')) %>%
  # only records that have accepted taxonomy for now
  #filter(taxonomicStatus == 'accepted') %>%
  #select only useful columns for now
  select(acceptedNameUsage, taxonomicStatus, species, subspecies, taxonRank,
         class, order, family, genus, 
         is_marine, is_brackish, is_freshwater,
         taxonID, ncbi_id, region) %>%
  # just in case
  distinct() %>%
  # rank to lowercase to match other databases
  mutate(taxonRank = tolower(taxonRank)) %>%
  # standardize column names
  janitor::clean_names() %>%
  mutate(verbatim_name = coalesce(subspecies, species)) %>%
  relocate(verbatim_name, .after = 'accepted_name_usage') %>%
  rename(accepted_name = accepted_name_usage)
  



write_csv(nep_fish, here(splists, '20230501_obis-fish_nep.csv'))
```

### Filter to Fishbase-validated list

Also make sure they all have NCBI ids!

```{r}
# use the accepted name to get a species list from fishbase
obisfish_base <- nep_fish %>%
  mutate(fishbase_name = validate_names(species_list = .$accepted_name))

# make a search for distribution() from rfishbase
obisfish_base_search <- obisfish_base %>%
  filter(!(is.na(fishbase_name))) %>%
  select(fishbase_name) %>%
  distinct()

# get distribution for fishbase names
dist_fish <- distribution(species_list = obisfish_base_search$fishbase_name, server = 'fishbase')

# filter for species by FAO region
# this should help account for errors in OBIS records
fishbase_in_FAO <- dist_fish %>%
  filter(FAO %in% 'Pacific, Northeast' |
           FAO %in% 'Pacific, Eastern Central' |
           FAO %in% 'America, North - Inland waters') %>%
  select(c('Species')) %>%
  distinct()

# filter the obis results by FAO region
obisfish_base <- obisfish_base %>%
  filter(fishbase_name %in% fishbase_in_FAO$Species)

# add NCBI IDs to any missing
ncbi_add <- obisfish_base %>%
  filter(is.na(ncbi_id)) %>%
  select(accepted_name) %>%
  distinct()

# right now I'm having issues with NCBI so I broke it up

ncbi_add1 <- ncbi_add[1:300,]
ncbi_add2 <- ncbi_add[301:591,]



# use gbif/ncbi to resolve the initial round of dirty names
ncbi_missing <- get_ids_(ncbi_add1$accepted_name, db = "ncbi")
ncbi_missing2 <- get_ids_(ncbi_add2$accepted_name, db = "ncbi")


# results to dataframe
ncbi_ids2 <- tibble(taxa_query = names(ncbi_missing$ncbi),
                      ncbi = ncbi_missing$ncbi) %>%
  # remove from lists
  unnest_longer(ncbi, keep_empty = TRUE) %>%
  # remove from next layer of lists
  unpack(cols = everything(), names_sep = "_")

# results to dataframe
ncbi_ids3 <- tibble(taxa_query = names(ncbi_missing2$ncbi),
                      ncbi = ncbi_missing2$ncbi) %>%
  # remove from lists
  unnest_longer(ncbi, keep_empty = TRUE) %>%
  # remove from next layer of lists
  unpack(cols = everything(), names_sep = "_") 


# all unique results
ncbi_ids <- ncbi_ids2 %>%
  full_join(ncbi_ids3) %>%
  distinct() %>%
  select(taxa_query, ncbi_uid) %>%
  rename(accepted_name = taxa_query)

# add new IDs to file
obisfish_base <- obisfish_base %>%
  left_join(ncbi_ids) %>%
  mutate(ncbi_id = coalesce(ncbi_id, as.numeric(ncbi_uid))) %>%
  select(!(c(ncbi_uid)))




write.csv(obisfish_base, here(splists, "20230504_obis-fish_fishbase_nep.csv"))
```

Problems for later
```{r}
no_ncbi <- obisfish_base %>%
  filter(is.na(ncbi_id)) %>%
  select(!(c(region))) %>%
  distinct()


write.csv(no_ncbi, here(splists, "20230504_obis-fish_no-ncbi.csv"))


no_ncbi_plot <- no_ncbi %>%
  select(accepted_name, family) %>%
  count(family)

# visualize missing

ggplot(no_ncbi_plot, 
       aes(x = reorder(family, -n), y = n)) +
  #geom_col(fill = 'darkslategray3') +
  geom_col(fill = 'darkslategray4') +
  labs(y = '# Species', 
       title = 'Families with species missing NCBI UIDs',
       x = NULL) + 
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 9),
        axis.text.y = element_text(size = 10),
        axis.title.y = element_text(size = 10))

ggsave(here('output', 'region_comparison', '20230504_familes-not-in-ncbi.png'), 
       last_plot(), 
       device = 'png',
       width = 17, height = 5,
       units = 'in')


no_ncbi_plot10 <- no_ncbi_plot %>%
  filter(n >= 10)

ggplot(no_ncbi_plot10, 
       aes(x = reorder(family, -n), y = n)) +
  #geom_col(fill = 'darkslategray3') +
  geom_col(fill = 'darkslategray4') +
  labs(y = '# Species', 
       title = 'Families with species missing NCBI UIDs',
       x = NULL) + 
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 9),
        axis.text.y = element_text(size = 10),
        axis.title.y = element_text(size = 10))

ggsave(here('output', 'region_comparison', '20230504_familes-not-in-ncbi_10.png'), 
       last_plot(), 
       device = 'png',
       width = 6, height = 5,
       units = 'in')
```


## Invertebrates

```{r invertebrates filter taxonomy}
obisinverts <- obisdat %>%
  # only species-level records
  filter(taxonRank %in% c('Species', 'Subspecies')) %>%
  # only animals
  filter(kingdom == 'Animalia') %>%
  # only not-vertebrates, 
  # but includes non-vertebrate chordates
  ## c() will also include any NAs in subphylum
  filter(!(subphylum %in% c('Vertebrata'))) %>%
  # only records that have accepted taxonomy for now
  filter(taxonomicStatus == 'accepted') %>%
  #select only useful columns for now
  select(species, subspecies, taxonRank,
         phylum, class, order, family, genus, 
         is_marine, is_brackish, is_freshwater,
         speciesid, taxonID, bold_id, ncbi_id) %>%
  # just in case
  distinct() %>%
  # rank to lowercase to match other databases
  mutate(taxonRank = tolower(taxonRank)) %>%
  # standardize column names
  janitor::clean_names()


write_csv(obisinverts, here(splists, '20230329_obis-invertebrates.csv'))

```
