==> scripts/old_version/fromgapp/03_species-in-region.Rmd <==
---
title: "03_species-in-region"
output: html_document
date: "2023-04-14"
---

This script takes the region lists from 01 with cleaned species lists to compare.

It will generate a column for in-region yes/no/na for the list. For the most accurate results, use WoRMS taxonomy to compare.


```{r setup, include=FALSE}
library(tidyverse)
library(arsenal)
library(eulerr)
library(gt)
library(gtExtras)
library(here)

splists <- here('processeddata', 'species_lists')
compare_out <- here('output', 'region_comparison')
```


```{r load-in}
# fish
region_fish <- read_csv(here(splists, 'region', '20230501_obis-fish_fishbase_nep.csv'))[-1]
compare_fish500 <- read_csv(here(splists, 'from_data', '20230501_gapp-1b-pacific_fish-top500-worrms.csv'))
compare_fishtully <- read_csv(here(splists, 'from_data', '20230323_gapp-3-tully_12sfish-sp.csv'))


# inverts
#region_inverts <- read_csv(here(splists, 'region', '20230412_obis-invertebrates_nep.csv'))

```
# Fish


## Top 500
Looking at hte top 500 blast hits
```{r fish top 500 eulerr}
# make list of fish in region
region_fish_sp <- region_fish %>%
  select(accepted_name) %>%
  distinct() %>%
  rename(compare_name = accepted_name)

# fish to compare from top 500
compare_fish500_sp <- compare_fish500 %>%
  # only ones that werent coalesced from verbatim_name
  filter(!(is.na(worms_name) & is.na(gbif_name))) %>%
  mutate(compare_name = found_taxa) %>%
  select(compare_name) %>%
  distinct()

venn_region_edna <- list('Region' = region_fish_sp$compare_name,
                         'eDNA, top 500' = compare_fish500_sp$compare_name
                         )

# run eulerr
euler_region_edna <- euler(venn_region_edna)

# plot output
plot(euler_region_edna,
     quantities = list(type = 'counts', cex = 1, fontface = 2),
     fills = list(fill = c('#a47148','#688ab5'), alpha = .75),
     adjust_labels = TRUE,
     main = 'Unique species')



# save
png(file = here(compare_out, 
                '20220501_gapp-1b-pacific_fish-top500_region-venn.png'), 
    width = 600, height = 400, units = 'px')
    plot(euler_region_edna,
     quantities = list(type = 'counts', cex = 1, fontface = 2),
     fills = list(fill = c('#a47148','#688ab5'), alpha = .75),
     adjust_labels = TRUE,
     main = 'Unique species')
    dev.off()
```

```{r fish top 500 arsenal}

# arsenal!
# in output, first dataframe is X, second is Y
region_500_comp <- comparedf(compare_fish500_sp, region_fish_sp, by = 'compare_name')

#extract differences
region_500_diff <- diffs(region_500_comp, what = c('observations'))

region_500_diff <- region_500_diff %>%
    mutate(version = recode(version, x = 'eDNA, top 500',
                            y = 'Region List')) %>%
  select(!(observation)) 

top500_not_in <- region_500_diff %>%
  filter(version == 'eDNA, top 500') %>%
  rename(source = version,
         found_taxa = compare_name)

# add column to top 500 for in-region
top500_update <- compare_fish500 %>%
  mutate(compare_name = found_taxa) %>%
  mutate(in_region = ifelse(compare_name %in% compare_fish500_sp$compare_name & 
                              compare_name %in% top500_not_in$compare_name, "no", 
                            ifelse(compare_name %in% compare_fish500_sp$compare_name & 
                                     !(compare_name %in% top500_not_in$compare_name), 'yes',
                                   NA_character_)))

write.csv(top500_update, here(splists, 'comparisons', '20230501_gapp-1b-pacific_fish-top500_compared-region.csv'))
write.csv(top500_not_in, here(splists, 'comparisons', '20230501_gapp-1b-pacific_fish-top500_out-of-region.csv'))


```


## Tully


```{r fish tully eulerr}
# tully is already a list of species
# region species list made in compare top 500

compare_fishtully <- compare_fishtully %>%
  rename(compare_name = species)


venn_region_tully <- list('Region' = region_fish_sp$compare_name,
                         'Tully' = compare_fishtully$compare_name
                         )

# run eulerr
euler_region_tully <- euler(venn_region_tully)

# plot output
plot(euler_region_tully,
     quantities = list(type = 'counts', cex = 1, fontface = 2),
     fills = list(fill = c('#a47148','#688ab5'), alpha = .75),
     adjust_labels = TRUE,
     main = 'Unique species')



# save
png(file = here(compare_out, 
                '20230501_gapp-3-tully_fish-mifish_region-venn.png'), 
    width = 600, height = 400, units = 'px')
    plot(euler_region_tully,
     quantities = list(type = 'counts', cex = 1, fontface = 2),
     fills = list(fill = c('#a47148','#688ab5'), alpha = .75),
     adjust_labels = TRUE,
     main = 'Unique species')
    dev.off()
```



```{r fish tully arsenal}


# arsenal!
# in output, first dataframe is X, second is Y
region_tully_comp <- comparedf(compare_fishtully, region_fish_sp, by = 'compare_name')

#extract differences
region_tully_diff <- diffs(region_tully_comp, what = c('observations'))

region_tully_diff <- region_tully_diff %>%
    mutate(version = recode(version, x = 'eDNA, Tully',
                            y = 'Region List')) %>%
  select(!(observation)) 

tully_not_in <- region_tully_diff %>%
  filter(version == 'eDNA, Tully')%>%
  rename(source = version,
         found_taxa = compare_name)

# add column to top 500 for in-region
tully_update <- compare_fishtully %>%
  mutate(in_region = ifelse(compare_name %in% compare_fishtully$compare_name & 
                              compare_name %in% tully_not_in$compare_name, "no", 
                            ifelse(compare_name %in% compare_fishtully$compare_name & 
                                     !(compare_name %in% tully_not_in$compare_name), 'yes',
                                   NA_character_)))

write.csv(tully_update, here(splists, 'comparisons', '20230501_gapp-3-tully_fish-mifish_compared-region.csv'))
write.csv(tully_not_in, here(splists, 'comparisons', '20230501_gapp-3-tully_fish-mifish_out-of-region.csv'))


```




# scratch

answer question of how many in region

```{r}


# maybe
# make summary stat table
region_500_stats <- region_500_diff %>%
  full_join(region_fish_sp) %>%
  # label ones found
  mutate(version = ifelse(is.na(version), 'both', 
                          ifelse(version == 'eDNA, top 500', 'top500', 'region'))) %>%
  group_by(version) %>%
  count() %>%
  ungroup() %>%
  pivot_wider(names_from = version, 
              values_from = n) %>%
  mutate(total_region = sum(region, both),
         total_top500 = sum(top500, both)) %>%
  mutate(percent_region_found = both/total_region) %>%
  mutate(percent_top500_inregion = both/total_top500)


# make summary stat table
region_500_stats2 <- region_500_diff %>%
  full_join(region_fish_sp) %>%
  # label ones found
  mutate(version = ifelse(is.na(version), 'both', 
                          ifelse(version == 'eDNA, top 500', 'top500', 'region'))) %>%
  group_by(version) %>%
  count() %>%
  ungroup() %>%
  mutate(total = n + n[version == 'both']) %>%
  mutate(percent_found = n[version == 'both']/total) #%>%
  mutate(percent_top500_inregion = both/total_top500)





```

```{r}
region_fish_sp2 <- region_fish %>%
  select(!(c(region))) %>%
  distinct(accepted_name, .keep_all = TRUE) 
```
